Replace TaskManager with TaskServices/TaskWorkflowHelper.
