+++ mdw-common/src/com/centurylink/mdw/model/task/TaskAction.java
@@ -47,7 +47,6 @@ public class TaskAction implements Serializable, Jsonable, Comparable<TaskAction
     public static final String ABORT = "Abort";
     public static final String WORK = "Work";
     public static final String SAVE = "Save";
-    public static final String CALLBACK = "Callback";    // for making a request/response call to the engine
 
     public static final String[] STANDARD_ACTIONS = { CREATE, ASSIGN, CLAIM, RELEASE, CANCEL, COMPLETE, RETRY, FORWARD, ABORT, WORK, SAVE };
