+++ mdw-services/src/com/centurylink/mdw/services/ServiceLocator.java
@@ -22,7 +22,6 @@ import com.centurylink.mdw.services.project.CollaborationServicesImpl;
 import com.centurylink.mdw.services.project.SolutionServicesImpl;
 import com.centurylink.mdw.services.request.RequestServicesImpl;
 import com.centurylink.mdw.services.system.SystemServicesImpl;
-import com.centurylink.mdw.services.task.TaskManagerBean;
 import com.centurylink.mdw.services.task.TaskServicesImpl;
 import com.centurylink.mdw.services.test.TestingServicesImpl;
 import com.centurylink.mdw.services.user.UserManagerBean;
@@ -39,22 +38,12 @@ public class ServiceLocator {
         return new UserManagerBean();
     }
 
-    public static TaskManager getTaskManager() {
-        return new TaskManagerBean();
-    }
-
     public static EventManager getEventManager() {
         return new EventManagerBean();
     }
 
     public static TaskServices getTaskServices() {
-        // TODO use ServiceRegistry
-        try {
-            return new TaskServicesImpl();
-        }
-        catch (Exception ex) {
-            throw new RuntimeException(ex.getMessage(), ex);
-        }
+        return new TaskServicesImpl();
     }
 
     public static ProcessServices getProcessServices() {
