+++ mdw-workflow/src/com/centurylink/mdw/workflow/task/strategy/ShortestQueueAutoAssignStrategy.java
@@ -15,7 +15,6 @@
  */
 package com.centurylink.mdw.workflow.task.strategy;
 
-import java.util.HashMap;
 import java.util.List;
 
 import com.centurylink.mdw.common.service.Query;
@@ -24,7 +23,6 @@ import com.centurylink.mdw.model.user.User;
 import com.centurylink.mdw.observer.ObserverException;
 import com.centurylink.mdw.observer.task.AutoAssignStrategy;
 import com.centurylink.mdw.services.ServiceLocator;
-import com.centurylink.mdw.services.TaskManager;
 import com.centurylink.mdw.services.UserManager;
 import com.centurylink.mdw.util.log.LoggerUtil;
 import com.centurylink.mdw.util.log.StandardLogger;
@@ -33,10 +31,9 @@ public class ShortestQueueAutoAssignStrategy implements AutoAssignStrategy {
 
     private static StandardLogger logger = LoggerUtil.getStandardLogger();
 
-    public User selectAssignee(TaskInstance taskInstanceVO) throws ObserverException {
+    public User selectAssignee(TaskInstance taskInstance) throws ObserverException {
         try {
-            TaskManager taskManager = ServiceLocator.getTaskManager();
-            List<String> groups = taskManager.getGroupsForTaskInstance(taskInstanceVO);
+            List<String> groups = ServiceLocator.getTaskServices().getGroupsForTaskInstance(taskInstance);
             UserManager userManager = ServiceLocator.getUserManager();
             User[] taskUsers = userManager.getUsersForGroups(groups.toArray(new String[groups.size()]));
             if (taskUsers == null || taskUsers.length == 0) {
